Single outcome mixin and planning alterations
refine_with_trailing_move() is now provided by the single outcome
Mixin. This is clearer, and avoids us to raise some Exceptions,
although we could make a variant for
Operations with several outcomes.

alter_destination() can now cope with Operation with
several outcomes. This avoids the [0] and wasn't harder to do.

Finally, doc is updated (and a missing crossref fixed)

In the future, a new mixin for Operations that manage their
outcomes destination could also be introduced.
