+++ anyblok_wms_base/core/operation/single_outcome.py
@@ -10,6 +10,9 @@
 """
 
 from anyblok import Declarations
+from anyblok_wms_base.exceptions import (
+    OperationError,
+    )
 
 Mixin = Declarations.Mixin
 Model = Declarations.Model
@@ -29,3 +32,49 @@ class WmsSingleOutcomeOperation:
         """
         Avatar = self.registry.Wms.PhysObj.Avatar
         return Avatar.query().filter_by(outcome_of=self).one()
+
+    def refine_with_trailing_move(self, stopover):
+        """Split the current Operation in two, the last one being a Move
+
+        This is for Operations that are responsible for the location of
+        their outcome (see the :attr:`destination_field
+        <anyblok_wms_base.core.operation.base.Operation.destination_field>`
+        class attribute)
+
+        :param stopover: this is the location of the intermediate Avatar
+                         that's been introduced (starting point of the Move).
+        :returns: the new Move instance
+
+        This doesn't change anything for the followers of the current
+        Operation, and in fact, it is guaranteed that their inputs are
+        untouched by this method.
+
+        Example use case: Rather than planning an Arrival followed by a Move to
+        stock location, One may wish to just plan an Arrival into some
+        the final stock destination, and later on, refine
+        this as an Arrival in a landing area, followed by a Move to the stock
+        destination. This is especially useful if the landing area can't be
+        determined at the time of the original planning, or simply to follow
+        the general principle of sober planning.
+        """
+        self.check_alterable()
+        field = self.destination_field
+        if field is None:
+            raise OperationError(
+                self,
+                "Can't refine {op} with a trailing move, because it's "
+                "not responsible for the location of its outcomes",
+                op=self)
+        setattr(self, field, stopover)
+
+        outcome = self.outcome
+        new_outcome = self.registry.Wms.PhysObj.Avatar.insert(
+            location=stopover,
+            outcome_of=self,
+            state='future',
+            dt_from=self.dt_execution,
+            # copied fields:
+            dt_until=outcome.dt_until,
+            obj=outcome.obj)
+        return self.registry.Wms.Operation.Move.plan_for_outcomes(
+            (new_outcome, ), (outcome, ), dt_execution=self.dt_execution)
